COH-23797 CPP: GitHub build of unit-test-lang is failing due to link error: missing powf
[git-p4: depot-paths = "//dev/release.cpp/coherence-cpp-v14.1.1.0/": change = 87319]

+++ tests/unit/coherence/lang/PrimitiveTest.hpp
@@ -11,7 +11,7 @@
 #include <iomanip>
 #include <iostream>
 #include <limits>
-#include <cmath>
+#include <math.h> // cmath has linker errors on some machines (missing pow functions)
 #include <sstream>
 
 using namespace coherence::lang;
@@ -148,7 +148,7 @@ class PrimitiveTest : public CxxTest::TestSuite
                 return true;
                 }
             // display the difference
-            float32_t flEpsilon = 1.0f / pow(10.0f, (float32_t) nPrecision);
+            float64_t flEpsilon = 1.0 / pow((float32_t) 10, nPrecision); // 64 bit float for epsilon to avoid loss of precision warnings on VS2019
             std::cout << std::setprecision(std::numeric_limits<float32_t>::digits10 + 2)
                       << "\nflVal1=" << flVal1 << ", flVal2=" << flVal2 << ", flDiff="
                       << flDiff << ", flEpsilon=" << flEpsilon << std::endl;
